separate user and amdin system

+++ routes/web.php
@@ -28,19 +28,19 @@ Route::middleware('admin')->prefix('admin')->group(function () {
     })->name('admin.index');
     // Route::resource('categories', CategoryController::class);
     Route::prefix('categories')->group(function () {
-        Route::get('/', [App\Http\Controllers\CategoryController::class, 'index'])->name('categories.index');
-        Route::post('/', [App\Http\Controllers\CategoryController::class, 'store'])->name('categories.store');
-        Route::delete('/{category}', [CategoryController::class, 'destroy'])->name('categories.destroy');
+        Route::get('/', [App\Http\Controllers\admin\CategoryController::class, 'index'])->name('categories.index');
+        Route::post('/', [App\Http\Controllers\admin\CategoryController::class, 'store'])->name('categories.store');
+        Route::delete('/{category}', [App\Http\Controllers\admin\CategoryController::class, 'destroy'])->name('categories.destroy');
     });
     Route::prefix('users')->group(function () {
-        Route::get('/', [App\Http\Controllers\UserController::class, 'index'])->name('users.index');
-        Route::get('/changeRole{user}', [App\Http\Controllers\UserController::class, 'changeRole'])->name('users.changeRole');
-        Route::get('/changeStatus{user}', [App\Http\Controllers\UserController::class, 'changeStatus'])->name('users.changeStatus');
-        Route::delete('/{user}', [App\Http\Controllers\UserController::class, 'destroy'])->name('users.destroy');
+        Route::get('/', [App\Http\Controllers\admin\UserController::class, 'index'])->name('users.index');
+        Route::get('/changeRole{user}', [App\Http\Controllers\admin\UserController::class, 'changeRole'])->name('users.changeRole');
+        Route::get('/changeStatus{user}', [App\Http\Controllers\admin\UserController::class, 'changeStatus'])->name('users.changeStatus');
+        Route::delete('/{user}', [App\Http\Controllers\admin\UserController::class, 'destroy'])->name('users.destroy');
     });
     Route::prefix('products')->group(function () {
-        Route::resource('products', App\Http\Controllers\ProductController::class);
-        Route::get('/changeStatus{product}', [App\Http\Controllers\ProductController::class, 'changeStatus'])->name('products.changeStatus');
+        Route::resource('products', App\Http\Controllers\admin\ProductController::class);
+        Route::get('/changeStatus{product}', [App\Http\Controllers\admin\ProductController::class, 'changeStatus'])->name('products.changeStatus');
     });
 
 });
